Pastebin Analog/pyproject.toml
[tool.poetry]
name = "pastebin"
version = "0.1.0"
description = "Pastebin Analog"
authors = ["Madjar Ivan <evanmadjar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
django-environ = "^0.10.0"
django-modeltranslation = "^0.18.9"
django-jazzmin = "^2.6.0"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.5"
django-cors-headers = "^3.14.0"
pillow = "^9.4.0"
mixer = "^7.2.2"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
